fully completed route to upload photo with post. also able to fully retrieve and display photos. Made image schema with relation to post
