+++ server.js
@@ -8,22 +8,8 @@ import likeRoutes from './controllers/likes.js';
 import profileRoutes from './controllers/profile.js';
 import authRequired from './middleware/authRequired.js';
 import { register, login, logout } from './controllers/auth.js';
-// import fileUpload from 'express-fileupload'
 import multer from 'multer';
-
-// let upload = multer({ dest: 'image' });
 import path from 'path';
-
-const upload = multer({
-	storage: multer.diskStorage({
-		destination: function (req, file, cb) {
-			cb(null, 'image/');
-		},
-		filename: function (req, file, cb) {
-			cb(null, new Date().valueOf() + '_' + file.originalname);
-		},
-	}),
-});
 
 const db = new prisma.PrismaClient({
 	log: ['info', 'warn'],
@@ -36,7 +22,6 @@ const port = 4000;
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cors());
-// app.use(fileUpload());
 
 // Controller(s)
 app.use('/users', authRequired, userRoutes);
@@ -48,14 +33,7 @@ app.use('/comments', authRequired, commentRoutes);
 app.use('/likes', authRequired, likeRoutes);
 app.use('/profile', authRequired, profileRoutes);
 
-app.post('/image', upload.single('image'), async (req, res) => {
-	const image = await db.image.create({
-		data: { ...req.file },
-	});
-	res.json({ image });
-});
-
-app.get('/image/:filename', (req, res) => {
+app.get('/image/:filename', authRequired, async (req, res) => {
 	const { filename } = req.params;
 	const dirname = path.resolve();
 	const fullfilepath = path.join(dirname, 'image/' + filename);
